brainbuddy_AI/requirements.txt
absl-py==2.3.1
ace_tools==0.0
alembic==1.16.4
attrs==25.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.59.0
fsspec==2025.7.0
git-filter-repo==2.47.0
greenlet==3.2.3
grpcio==1.73.1
idna==3.10
imageio==2.37.0
intel-openmp==2021.4.0
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
lightgbm==4.6.0
llvmlite==0.44.0
Mako==1.3.10
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.3
mediapipe==0.10.9
mkl==2021.4.0
ml_dtypes==0.5.1
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
optuna==4.4.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==3.20.3
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pynndescent==0.5.13
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
sentencepiece==0.2.0
shap==0.48.0
six==1.17.0
slicer==0.0.8
sounddevice==0.5.2
SQLAlchemy==2.0.41
sympy==1.14.0
tbb==2021.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.5.10
tomli==2.2.1
torch @ file:///C:/Users/user/Downloads/torch-2.3.1%2Bcu121-cp310-cp310-win_amd64.whl
torchvision @ file:///C:/Users/user/Downloads/torchvision-0.18.1%2Bcu121-cp310-cp310-win_amd64.whl
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.168
ultralytics-thop==2.0.14
umap-learn==0.5.9.post2
urllib3==2.5.0
Werkzeug==3.1.3
